[567249] Added icons, changed editor
