+++ core/plugins/org.polarsys.capella.core.platform.eclipse.ui.trace/META-INF/MANIFEST.MF
@@ -8,7 +8,8 @@ Require-Bundle: org.eclipse.ui,
  org.eclipse.core.runtime,
  org.polarsys.capella.common.ui.services,
  org.polarsys.capella.core.ui.toolkit,
- org.polarsys.capella.common.ui
+ org.polarsys.capella.common.ui,
+ org.polarsys.capella.core.ui.properties
 Bundle-ActivationPolicy: lazy
 Export-Package: org.polarsys.capella.core.platform.eclipse.capella.ui.trace,
  org.polarsys.capella.core.platform.eclipse.capella.ui.trace.commands,
