+++ core/plugins/org.polarsys.capella.core.platform.eclipse.ui.trace/plugin.properties
@@ -1,5 +1,5 @@
 # =============================================================================
-# Copyright (c) 2006, 2015 THALES GLOBAL SERVICES.
+# Copyright (c) 2006, 2020 THALES GLOBAL SERVICES.
 # 
 # This program and the accompanying materials are made available under the
 # terms of the Eclipse Public License 2.0 which is available at
@@ -13,3 +13,4 @@
 #Properties file for org.polarsys.capella.core.platform.eclipse.ui.trace
 providerName = Eclipse.org
 pluginName = MDTrace Eclipse UI For Capella
+TraceManager=Trace Manager
