+++ core/plugins/org.polarsys.capella.core.platform.eclipse.ui.trace/src/org/polarsys/capella/core/platform/eclipse/capella/ui/trace/messages/messages.properties
@@ -1,5 +1,5 @@
 #===============================================================================
-# Copyright (c) 2006, 2018 THALES GLOBAL SERVICES.
+# Copyright (c) 2006, 2020 THALES GLOBAL SERVICES.
 # 
 # This program and the accompanying materials are made available under the
 # terms of the Eclipse Public License 2.0 which is available at
@@ -50,9 +50,6 @@ MDTrace.window_title=Traceability Management Wizard
 MdeElementTreeSelectionDialog.filter_char=? = any character, * = any string
 MdeElementTreeSelectionDialog.group_text=Select a name to find
 TraceTreeViewer.void_trace_error=Void traces are not allowed, select a model element to trace.
-
-TracePageContentProvider.page.title=Capella trace
-TracePageContentProvider.page.description=This editor displays Capella trace.
 
 # The tab strings.
 TraceGeneralTabDescriptor.tab.title=&General
