+++ core/plugins/org.polarsys.capella.core.platform.sirius.ui.actions/META-INF/MANIFEST.MF
@@ -12,7 +12,6 @@ Require-Bundle: com.google.guava,
  org.eclipse.sirius.ui;visibility:=reexport,
  org.polarsys.capella.common.tools.report.appenders.reportlogview;visibility:=reexport,
  org.polarsys.capella.common.ui.services;visibility:=reexport,
- org.polarsys.capella.core.platform.eclipse.ui.trace,
  org.polarsys.capella.core.preferences,
  org.polarsys.capella.core.ui.toolkit,
  org.eclipse.sirius.diagram.ui,
